fix(aside): Clip the dot that hides the corner
A background was applied to the corner icon of the aside to give the cool
effect that the corner has been cut off. But the issue with that is that
the background goes out quite far beyond the icon, which can potentially
overlap with stuff behind it.

By adding a clip-path, the overlapping background now only goes over the
corner of the aside (lower right 25% of the icon) rather than around the
whole icon.