feat(cooking): Show a different card for recipe list on mobile
For some reason I can't figure out, both the CSS hover and onTouchStart
events are only working for the first recipe card that is being touched
now that the app is built with Astro.

I can't even come up with a theory as to why that is happening, so
I just made a different card to show in that case that doesn't need
hover to show the title like the old one did.