chore: Use deafult import from react-color
If I try to use the named export from react-color, the build fails because it thinks that react-color is using CommonJS (when it apperas to support es modules).

If I use DefaultExport.ChromePicker I get an error because ChromePicker doesn't exist.

Luckily, The ChromePicker is actually the default export of the
entire library, which seems to work fine. Stupid module shit.